refactor: split isolation.rs into files
This is a form of preparation for the introduction of additional
security-related measures, which may be best to implement in separate
files.
